v1.2.4
+ Added COPYING license file (GNU GPL 3.0)
+ Added hovering over components now has a satisfying light up animation
* Improved menubar, now supports dividers and element visibility
* Changed color theme (now darker, and window border lights up when focused)
* Improved ColorTheme.ini, much fewer elements, and they are easily recognized by name
* Improved exit dialog to show for all open files when closing program
* Fixed problem with closing a model would disable editting for other open models

+++ include/engine/utils/variable/datatype/Color.h
@@ -7,11 +7,11 @@
 struct Color {
 private:
 	GLfloat sub(GLfloat a, GLfloat b) {
-		return GLfloat(fmax(a - b, 0));
+		return GLfloat(fmax(static_cast<double>(a) - b, 0));
 	}
 
 	GLfloat add(GLfloat a, GLfloat b) {
-		return GLfloat(fmin(a + b, 1.f));
+		return GLfloat(fmin(static_cast<double>(a) + b, 1.f));
 	}
 public:
 	GLfloat r, g, b, a;
